LangChain/WithHuggingFace/requirements.txt
python-dotenv
langchain-huggingface
transformers
datasets